agents/quantum_training_agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quantum_training_agent"
version = "0.1.0"
authors = [
    { name="Quantum Training Team", email="quantum@example.com" },
]
description = "A quantum-enhanced training agent for language models"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
]
dependencies = [
    "numpy>=1.21.0",
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "unsloth>=0.3.0",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.41.0",
    "peft>=0.7.0",
    "azure-quantum>=1.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.92.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0"
]

[project.urls]
"Homepage" = "https://github.com/username/quantum_training_agent"
"Bug Tracker" = "https://github.com/username/quantum_training_agent/issues"